Add core error handling and home feature entities and repository
This commit introduces the foundational error handling mechanism by adding the `Failure` class, which will be used to encapsulate error messages throughout the application. Additionally, the commit sets up the domain layer for the home feature by defining the `BookEntity` class, which represents the essential attributes of a book. The `HomeRepo` abstract class is also introduced to outline the repository interface for fetching featured and newest books, utilizing the `dartz` package for functional error handling with `Either`. Lastly, the `dartz` package is added to the project dependencies in `pubspec.yaml` and `pubspec.lock` to support the functional programming paradigm used in the repository implementation.