Add UseCase abstract class and FetchFeaturedBooksUseCase implementation
This commit introduces the `UseCase` abstract class, which serves as a base for all use cases in the application. The `UseCase` class defines a generic method `call` that returns a `Future` containing either a `Failure` or the expected result type. Additionally, a `NoParam` class is provided for use cases that do not require parameters.

The `FetchFeaturedBooksUseCase` class is implemented as a concrete use case that fetches a list of featured books. It extends the `UseCase` class and overrides the `call` method to delegate the fetching operation to the `HomeRepo` repository. This use case does not require any parameters, hence it uses the `NoParam` class.