Add FeaturedBooksCubit and State Classes
Created a new `FeaturedBooksCubit` class in the `home` feature to manage the state of featured books. This includes the initial state, loading state, failure state, and success state with a list of `BookEntity` objects. Also, added the `bloc` package to `pubspec.yaml` and updated `pubspec.lock` to reflect this dependency.